Separated logic at club, event services
1. The ClubAndEventService has been dropped, and its logic has been split into separate services.
2. Minor refactors at files that used the dropped service in (1), among others.
