+++ app/Http/Controllers/ClubController.php
@@ -4,63 +4,63 @@ namespace App\Http\Controllers;
 
 use App\Models\Club;
 use Illuminate\Http\Request;
-use App\Services\ClubAndEventService;
+use App\Services\ClubService;
 use Illuminate\Support\Facades\Storage;
 
 class ClubController extends Controller
 {
-    protected $clubAndEventService;
+    protected $clubService;
 
-    public function __construct(ClubAndEventService $clubAndEventService) {
-        $this->clubAndEventService = $clubAndEventService;
+    public function __construct(ClubService $clubService) {
+        $this->clubService = $clubService;
     }
 
     public function fetchClubsFinder(Request $request) {
-        return $this->clubAndEventService->prepareAndRenderClubsFinderView($request);
+        return $this->clubService->prepareAndRenderClubsFinderView($request);
     }
 
     public function clearFilterForGeneral() {
-        return $this->clubAndEventService->flushClubFilters('clubs-finder');
+        return $this->clubService->flushClubFilters('clubs-finder');
     }
-
+    
     public function fetchClubDetailsForGeneral(Request $request) {
         return view(
             'clubs-finder.general.view-club-details',
-            $this->clubAndEventService->prepareClubData($request->query('club_id'))
+            $this->clubService->prepareClubData($request->query('club_id'))
         );
     }
 
     public function fetchClubsManager(Request $request) {
-        return $this->clubAndEventService->prepareAndRenderClubsFinderView($request);
+        return $this->clubService->prepareAndRenderClubsFinderView($request);
     }
 
     public function clearFilterForManager() {
-        return $this->clubAndEventService->flushClubFilters('manage-clubs');
+        return $this->clubService->flushClubFilters('manage-clubs');
     }
 
     public function fetchClubDetailsForManager(Request $request) {
         return view(
             'clubs-finder.admin-manage.view-club-details',
-            $this->clubAndEventService->prepareClubData($request->query('club_id'))
+            $this->clubService->prepareClubData($request->query('club_id'))
         );
     }
 
     public function fetchAdminManagePage(Request $request) {
         return view(
             'clubs-finder.admin-manage.manage-club-details',
-            $this->clubAndEventService->prepareClubData($request->query('club_id'))
+            $this->clubService->prepareClubData($request->query('club_id'))
         );
     }
 
     public function fetchCommitteeManagePage(Request $request) {
         return view(
             'clubs-finder.committee-manage.manage-club-details',
-            $this->clubAndEventService->prepareClubData($request->query('club_id'))
+            $this->clubService->prepareClubData($request->query('club_id'))
         );
     }
 
     public function showClubInfoEditForAdmin(Request $request) {
-        $data = $this->clubAndEventService->prepareClubData($request->query('club_id'));
+        $data = $this->clubService->prepareClubData($request->query('club_id'));
 
         return view('clubs-finder.admin-manage.edit.club-info', [
             'club' => $data['club'],
@@ -68,7 +68,7 @@ class ClubController extends Controller
     }
 
     public function showClubInfoEditForCommittee(Request $request) {
-        $data = $this->clubAndEventService->prepareClubData($request->query('club_id'));
+        $data = $this->clubService->prepareClubData($request->query('club_id'));
 
         return view('clubs-finder.committee-manage.edit.club-info', [
             'club' => $data['club'],
@@ -76,7 +76,7 @@ class ClubController extends Controller
     }
 
     public function showClubImagesEditForAdmin(Request $request) {
-        $data = $this->clubAndEventService->prepareClubData($request->query('club_id'));
+        $data = $this->clubService->prepareClubData($request->query('club_id'));
 
         return view('clubs-finder.admin-manage.edit.images', [
             'club' => $data['club'],
@@ -84,7 +84,7 @@ class ClubController extends Controller
     }
 
     public function showClubImagesEditForCommittee(Request $request) {
-        $data = $this->clubAndEventService->prepareClubData($request->query('club_id'));
+        $data = $this->clubService->prepareClubData($request->query('club_id'));
 
         return view('clubs-finder.committee-manage.edit.images', [
             'club' => $data['club'],
